+++ apps/@sparrow-desktop/src/pages/dashboard-page/Dashboard.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
   import {
-    GlobalSearch,
     LoginBanner,
     LoginSignupConfirmation,
     SwitchWorkspace,
@@ -41,6 +40,7 @@
     SidebarItemPositionBaseEnum,
     SidebarItemIdEnum,
   } from "@sparrow/common/types/sidebar/sidebar-base";
+  import { GlobalSearch } from "@sparrow/common/features";
 
   const _viewModel = new DashboardViewModel();
   let userId;
