+++ packages/@sparrow-common/src/components/global-search/sub-components/types/types.ts
@@ -1,22 +1,2 @@
-export interface SearchSuggestion {
-  type:
-    | "workspace"
-    | "collection"
-    | "environment"
-    | "folder"
-    | "flow"
-    | "request";
-  label: string;
-  icon: string;
-}
-
-export interface RecentItem {
-  title: string;
-  workspace: string;
-  collection: string;
-  url: string;
-  icon?: string;
-  method?: string;
-}
