+++ packages/@sparrow-common/src/components/index.ts
@@ -8,5 +8,4 @@ export { default as PulseCircle } from "./pulse-circle/PulseCircle.svelte";
 export { default as WelcomePopUpWeb } from "./welcome-pop-up-web/WelcomePopUpWeb.svelte";
 export { default as DefaultHeader } from "./default-header/DefaultHeader.svelte";
 export { default as LaunchDesktop } from "./lanuch-desktop/LanuchDesktop.svelte";
-export { default as GlobalSearch } from "./global-search/GlobalSearch.svelte";
 export { default as SwitchWorkspace } from "./switch-workspace/SwitchWorkspace.svelte";
