+++ packages/@sparrow-common/src/features/global-search/images/Board.svelte
