+++ packages/@sparrow-common/src/features/global-search/images/Delete.svelte
