+++ packages/@sparrow-common/src/features/global-search/images/Get.svelte
