+++ packages/@sparrow-common/src/features/global-search/images/GraphQL.svelte
