+++ packages/@sparrow-common/src/features/global-search/images/Patch.svelte
