+++ packages/@sparrow-common/src/features/global-search/images/Post.svelte
