+++ packages/@sparrow-common/src/features/global-search/images/Put.svelte
