+++ packages/@sparrow-common/src/features/global-search/images/Search.png
