+++ packages/@sparrow-common/src/features/global-search/images/SearchIcon.svelte
