+++ packages/@sparrow-common/src/features/global-search/images/Socket.io.svelte
