+++ packages/@sparrow-common/src/features/global-search/images/WebSocket.svelte
