+++ packages/@sparrow-common/src/features/global-search/images/arrowSwap.svelte
