+++ packages/@sparrow-common/src/features/global-search/images/collectionStack.svelte
