+++ packages/@sparrow-common/src/features/global-search/images/envLayer.svelte
