+++ packages/@sparrow-common/src/features/global-search/images/flowIcon.svelte
