+++ packages/@sparrow-common/src/features/global-search/images/folderIcon.svelte
