+++ packages/@sparrow-common/src/features/global-search/images/index.ts
@@ -1,0 +1,19 @@
+export { default as RequestIcon } from "./arrowSwap.svelte";
+export { default as CollectionIcon } from "./collectionStack.svelte";
+export { default as crossIcon } from "./crossIcon.svg";
+export { default as deleteIcon } from "./Delete.svelte";
+export { default as EnvironmentIcon } from "./envLayer.svelte";
+export { default as FlowIcon } from "./flowIcon.svelte";
+export { default as FolderIcon } from "./folderIcon.svelte";
+export { default as getIcon } from "./Get.svelte";
+export { default as keyCommand } from "./keyCommand.png";
+export { default as patchIcon } from "./Patch.svelte";
+export { default as postIcon } from "./Post.svelte";
+export { default as putIcon } from "./Put.svelte";
+export { default as Search } from "./Search.png";
+export { default as searchIcon } from "./searchIcon.png";
+export { default as socketIoIcon } from "./Socket.io.svelte";
+export { default as webSocketIcon } from "./WebSocket.svelte";
+export { default as WorkspaceIcon } from "./Board.svelte";
+export { default as hexIcon } from "./GraphQL.svelte";
+export { default as SearchIcon } from "./SearchIcon.svelte";
