+++ packages/@sparrow-common/src/features/global-search/images/searchIcon.png
