+++ packages/@sparrow-common/src/features/global-search/index.ts
@@ -1,0 +1,1 @@
+export { default as GlobalSearch} from "./layout/GlobalSearch.svelte";
