+++ packages/@sparrow-common/src/features/index.ts
@@ -3,3 +3,4 @@ export * from "./app-updater";
 export * from "./download-app";
 export * from "./create-team";
 export * from "./sidebar";
+export * from "./global-search";
