+++ packages/@sparrow-common/src/images/index.ts
@@ -5,23 +5,4 @@ export { default as sparrowLogoBackground } from "./sparrowLogoBackground.svg";
 export { default as SparrowBackgroundV2 } from "./SparrowBackgroundV2.svg";
 export { default as WelcomeLogo } from "./welcome-logo.svelte";
 export { default as background } from "./background.png";
-export { default as RequestIcon } from "./arrowSwap.svelte";
-export { default as CollectionIcon } from "./collectionStack.svelte";
-export { default as crossIcon } from "./crossIcon.svg";
-export { default as deleteIcon } from "./Delete.svg";
-export { default as EnvironmentIcon } from "./envLayer.svelte";
-export { default as FlowIcon } from "./flowIcon.svelte";
-export { default as FolderIcon } from "./folderIcon.svelte";
-export { default as getIcon } from "./Get.svg";
-export { default as keyCommand } from "./keyCommand.png";
-export { default as patchIcon } from "./Patch.svg";
-export { default as postIcon } from "./Post.svg";
-export { default as putIcon } from "./Put.svg";
-export { default as Search } from "./Search.png";
-export { default as searchIcon } from "./searchIcon.png";
 export { default as Shape } from "./Shape.png";
-export { default as socketIoIcon } from "./Socket.io.svg";
-export { default as webSocketIcon } from "./WebSocket.svg";
-export { default as WorkspaceIcon } from "./Board.svelte";
-export { default as hexIcon } from "./GraphQL.svg";
-export { default as SearchIcon } from "./SearchIcon.svelte";
